Added tabs.

+++ DataPool.py
@@ -5,13 +5,13 @@ import logging
 
 class DataPool:
     def __init__(self, size: int):
-        self.record_buffers: List[List[int]] = [list() for _ in range(0, size)]
+        self.record_buffers: List[List[float]] = [list() for _ in range(0, size)]
         self.size: int = size
         self.label: str = "None"
-        self.circular_buffers: List[Deque[int]] = [deque(maxlen=3000) for _ in range(0, size)]
+        self.circular_buffers: List[Deque[float]] = [deque(maxlen=3000) for _ in range(0, size)]
         self.recording: bool = False
 
-    def append(self, pin: int, data: int):
+    def append(self, pin: int, data: float):
         try:
             if self.recording:
                 self.record_buffers[pin].append({"label": self.label, "data": data})
